feat: added lambdas delete patch and post fixed minor issues in other lambdas
