+++ src/lambda/delete-thread/index.py
@@ -1,16 +1,18 @@
 from boto3.dynamodb.conditions import Attr
+import boto3
 
 from utils import JsonPayloadBuilder, table, resp_handler
 
 
 @resp_handler
-def delete_thread(key, ts, uid):
+def delete_thread(board_id, thread_id, ts, uid):
     table.delete_item(
         Key={
-            "course_key": key,
+            "board_id": board_id,
             "created_at": ts
         },
-        ConditionExpression=Attr('uid').eq(uid)
+        ConditionExpression=Attr('uid').eq(
+            uid) & Attr('thread_id').eq(thread_id)
     )
 
     resp_body = JsonPayloadBuilder().add_status(
@@ -20,7 +22,8 @@ def delete_thread(key, ts, uid):
 
 def handler(event, context):
     params = {
-        "key": event["pathParameters"]["thread_id"],
+        "board_id": event["pathParameters"]["board_id"],
+        "thread_id": event["pathParameters"]["thread_id"],
         "ts": event["queryStringParameters"]["ts"],
         "uid": event['requestContext']['authorizer']['claims']['sub']
     }
