+++ src/lambda/patch-thread/index.py
@@ -1,2 +1,36 @@
+import json
+from datetime import datetime
+from utils import JsonPayloadBuilder
+from utils import resp_handler
+from utils import table
+
+
+@resp_handler
+def patch_thread(board_id, uid, thread):
+    dt_now = datetime.now().strftime('%Y-%m-%dT%H:%M:%S.%f')[:-3] + 'Z'
+    table.update_item(
+        Key={
+            "uid": uid
+        },
+        UpdateExpression='SET courses = :info, update_at = :ts',
+        ExpressionAtrributeValues={
+            ":info": [body['courses']],
+            ":ts": dt_now
+        }
+    )
+
+    body = JsonPayloadBuilder().add_status(
+        True).add_data(None).add_message('').compile()
+    return body
+
+
 def handler(event, context):
-    pass
+    req = json.loads(event['body'])
+    params = {
+        "board_id": event["pathParameters"]["board_id"],
+        "thread_id": event["pathParameters"]["thread_id"],
+        "uid": event['requestContext']['authorizer']['claims']['sub'],
+        "thread": req["data"]
+    }
+
+    return patch_thread(**params)
